Split Quartus types into separate file
