+++ hls4ml/backends/quartus/passes/transform_types.py
@@ -1,12 +1,10 @@
-from hls4ml.backends.fpga.fpga_types import (
-    ACTypeConverter,
-    HLSTypeConverter,
+from hls4ml.backends.fpga.fpga_types import ACTypeConverter, HLSTypeConverter, StaticWeightVariableConverter
+from hls4ml.backends.quartus.quartus_types import (
     QuartusArrayVariableConverter,
     QuartusInplaceArrayVariableConverter,
     QuartusInplaceStreamVariableConverter,
     QuartusStreamVariableConverter,
     QuartusStructMemberVariableConverter,
-    StaticWeightVariableConverter,
 )
 from hls4ml.model.optimizer import GlobalOptimizerPass
 from hls4ml.model.types import InplaceTensorVariable
